优化前端管理页面：隐藏ID列，优化表格列宽，提升用户体验
- 移除所有管理页面的ID列显示，提升安全性和用户体验
- 将固定width改为响应式min-width，适配不同屏幕尺寸
- 添加show-overflow-tooltip处理长文本显示
- 优化权限管理页面，移除parentId列，保留更有意义的parentName
- 添加代码质量分析规则文件
- 清理测试报告和临时文件

+++ deyoch-oa-backend/src/main/java/com/deyoch/controller/DeptController.java
@@ -105,7 +105,7 @@ public class DeptController {
     @PostMapping("/{id}/status")
     @PreAuthorize("hasAuthority('sys:dept:manage')")
     @Operation(summary = "更新部门状态", description = "根据部门ID更新部门状态")
-    public Result<Void> updateDeptStatus(@PathVariable @Parameter(description = "部门ID") Long id, @RequestParam @Parameter(description = "部门状态") Long status) {
+    public Result<Void> updateDeptStatus(@PathVariable @Parameter(description = "部门ID") Long id, @RequestParam @Parameter(description = "部门状态") Integer status) {
         return deptService.updateDeptStatus(id, status);
     }
 }
