+++ deyoch-oa-backend/src/main/java/com/deyoch/entity/DeyochPermission.java
@@ -22,8 +22,8 @@ public class DeyochPermission {
   private String path;
   private String component;
   private String icon;
-  private Long sort;
-  private Long status;
+  private Integer sort;
+  private Integer status;
   @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
   private LocalDateTime createdAt;
   @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
